+++ packages/api/models/index.js
@@ -3,11 +3,9 @@
 const AccessToken = require('./access-tokens');
 const Manager = require('./base');
 const ReconciliationReport = require('./reconciliation-reports');
-const Rule = require('./rules');
 
 module.exports = {
   AccessToken,
   ReconciliationReport,
-  Rule,
   Manager,
 };