+++ CHANGELOG.md
@@ -85,6 +85,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - **CUMULUS-2792**
     - Remove API Granule model and all related test code
     - Remove granule-csv endpoint
+  - **CUMULUS-2797**
+    - Remove API Rules model and all related test code
+    - Move rule helper functions to separate rulesHelpers file
   - **CUMULUS-2821**
     - Remove DynamoDB logic from `sfEventSqsToDbRecords` lambda
   - **CUMULUS-2856**